+++ p2panda/src/streams/sync_metrics.rs
@@ -17,13 +17,13 @@ type SessionId = u64;
 #[derive(Clone, Debug, Default)]
 pub struct Aggregator {
     /// Total number of running sync sessions for a topic.
-    running_sessions: u64,
+    running_sessions: u32,
 
     /// Total number of bytes sent across all topic sessions.
-    total_bytes_sent: u64,
+    total_bytes_sent: u32,
 
     /// Total number of received bytes across all topic sessions.
-    total_bytes_received: u64,
+    total_bytes_received: u32,
 
     /// Latest metrics for all sessions.
     session_metrics: HashMap<SessionId, Metrics>,
@@ -141,17 +141,17 @@ impl Aggregator {
     }
 
     /// Total running sessions for a topic.
-    pub fn running_sessions(&self) -> u64 {
+    pub fn running_sessions(&self) -> u32 {
         self.running_sessions
     }
 
     /// Total bytes sent on a topic.
-    pub fn total_bytes_sent(&self) -> u64 {
+    pub fn total_bytes_sent(&self) -> u32 {
         self.total_bytes_sent
     }
 
     /// Total bytes received on a topic.
-    pub fn total_bytes_received(&self) -> u64 {
+    pub fn total_bytes_received(&self) -> u32 {
         self.total_bytes_received
     }
 }
@@ -177,21 +177,21 @@ pub(crate) enum SyncEvent<E> {
     SyncStarted {
         remote: NodeId,
         session_id: u64,
-        incoming_operations: u64,
-        outgoing_operations: u64,
-        incoming_bytes: u64,
-        outgoing_bytes: u64,
-        topic_sessions: u64,
+        incoming_operations: u32,
+        outgoing_operations: u32,
+        incoming_bytes: u32,
+        outgoing_bytes: u32,
+        topic_sessions: u32,
     },
     SyncEnded {
         remote: NodeId,
         session_id: u64,
-        sent_operations: u64,
-        received_operations: u64,
-        sent_bytes: u64,
-        received_bytes: u64,
-        sent_bytes_topic_total: u64,
-        received_bytes_topic_total: u64,
+        sent_operations: u32,
+        received_operations: u32,
+        sent_bytes: u32,
+        received_bytes: u32,
+        sent_bytes_topic_total: u32,
+        received_bytes_topic_total: u32,
         error: Option<SyncError>,
     },
     OperationReceived {
